Flip Into<PrintItems> impls to From
This change flips the `impl Into<PrintItem> for _` to `impl From<_> for PrintItem`. The impact on the rest of the code, and the public facing API, is nil but the benefit is (from the rustdocs for [`std::convert::From`](https://doc.rust-lang.org/std/convert/trait.From.html)

> One should always prefer implementing From over Into because implementing From automatically provides one with an implementation of Into thanks to the blanket implementation in the standard library.

There's an additional impl I added which says to implement `From<Option<T>>` for all `T`s where there is a `impl From<T> for PrintItems`. I thought that this would be better than `impl From<Option<PrintItemPath>> ...`, and any additional `From<Option<_>>` that would be useful to add (as they're all added by this generic trait impl).