feat: add `allow-no-files` flag and config option
This option allows to exit `drpint` process with code `0` when
there is some file passed to input and this file is also added to
`excludes` property or cli arg.

For example, let's take `tsconfig.json`. I can mark it as excluded,
but when my pre-commit hook will pass this file to `dprint fmt`,
I got an error `No files found to format with the specified plugins`.
But when I pass `--allow-no-files` or set this option in config,
I can ran `dprint fmt` whout any issue.